requirements.txt
django
djangorestframework
psycopg2-binary
flake8
python-dotenv
gunicorn
requests
django-filter
